Added initial files.
These are the engine and tests for my previous projects. The projects were either ASCII projects or editted files directly. I've changed some things to simplify before bringing this over to GitHub. It's noticeable that the tests don't currently compile, as I was in the middle of messing with some things when I started the process of moving these over.